+++ finite-automata-designer/public/scripts/Shapes/Arrow.ts
@@ -14,8 +14,8 @@ import {Circle} from "./Circle";
 import { SelfArrow } from "./SelfArrow";
 import { EntryArrow } from "./EntryArrow";
 import { snapToPadding, nodeRadius, drawArrow, drawText, hitTargetPadding } from "./draw";
-import { ExportAsSVG } from "./exporting/ExportAsSVG";
-import { ExportAsLaTeX } from "./exporting/ExportAsLaTeX";
+import { ExportAsSVG } from "../exporting/ExportAsSVG";
+import { ExportAsLaTeX } from "../exporting/ExportAsLaTeX";
 
 export var arrows: (Arrow | SelfArrow | EntryArrow)[] = [];
